refactor: move tr_peer_info to peer-mgr.h
